Implementing Singleton in C#
- Why Singleton ?
This pattern involves a single class which is responsible to create an
object while making sure that only single object gets created.
- Advantage
The instantiation is not performed until an object asks for an instance;
this approach is referred to as lazy instantiation. Lazy instantiation
avoids instantiating unnecessary singletons when the application starts.
- Disadvantage
It is not safe for multithreaded environments. If separate threads of
execution enter the Instance property method at the same time, more that
one instance of the Singleton object may be created.